@@ -8,54 +8,20 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.5.4.RELEASE</version>
+		<version>2.7.15</version>
 	</parent>
 
 	<properties>
-		<apache.cxf.version>3.1.7</apache.cxf.version>
- 		<mysql.connector.version>5.1.40</mysql.connector.version>
- 		<postgresql.version>42.3.3</postgresql.version>
- 		<h2.version>1.4.192</h2.version>
- 		<gson.version>2.7</gson.version>
- 		<flyway.version>4.2.0</flyway.version>
-		<javax.mail-api.version>1.5.6</javax.mail-api.version>
-		<javax.mail.version>1.4.7</javax.mail.version>
+        <java.version>11</java.version>
 		<dbunit.version>2.5.3</dbunit.version>
 		<dbunit.spring.test.version>1.3.0</dbunit.spring.test.version>
 		<jjwt.version>0.6.0</jjwt.version>
 		<modelmapper.version>0.7.7</modelmapper.version>
 		<springfox.swagger2.version>2.7.0</springfox.swagger2.version>
 		<sendinblue.version>6.0.0</sendinblue.version>
-		<thymeleaf.version>3.0.2.RELEASE</thymeleaf.version>
 	</properties>
 
 	<dependencies>
-		<!-- API, java.xml.bind module -->
-		<dependency>
-			<groupId>jakarta.xml.bind</groupId>
-			<artifactId>jakarta.xml.bind-api</artifactId>
-			<version>2.3.2</version>
-		</dependency>
-
-		<!-- Runtime, com.sun.xml.bind module -->
-		<dependency>
-			<groupId>org.glassfish.jaxb</groupId>
-			<artifactId>jaxb-runtime</artifactId>
-			<version>2.3.2</version>
-		</dependency>
-
-		<!-- Java MailSender -->
-		<dependency>
-			<groupId>javax.mail</groupId>
-			<artifactId>javax.mail-api</artifactId>
-			<version>${javax.mail-api.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>javax.mail</groupId>
-			<artifactId>mail</artifactId>
-			<version>${javax.mail.version}</version>
-		</dependency>
-
 		<!-- SendInBlue MailSender -->
 		<dependency>
 			<groupId>com.sendinblue</groupId>
