@@ -15,35 +15,35 @@ import org.springframework.stereotype.Component;
 public class DecisionConverter {
 	private final VoteOptionRepository voteOptionRepository;
 
-	public final Converter<VoteDecision, VoteDecisionDTO> toDTO = new AbstractConverter<VoteDecision, VoteDecisionDTO>() {
-		@Override
-		protected VoteDecisionDTO convert(VoteDecision voteDecision) {
-			VoteDecisionDTO dtoDecision = new VoteDecisionDTO();
-			dtoDecision.setOptionId(voteDecision.getVoteOption().getId());
-			dtoDecision.setParticipantId(voteDecision.getVoteParticipant().getId());
-			
-			if (voteDecision.getDecision() != null) {
-				dtoDecision.setDecision(voteDecision.getDecision().getValue());
-			} else if (voteDecision.getParticipants() != null) {
-				dtoDecision.setParticipants(voteDecision.getParticipants());
-			}
-			return dtoDecision;
-		}
-	};
+	public final Converter<VoteDecision, VoteDecisionDTO> toDTO = new AbstractConverter<>() {
+        @Override
+        protected VoteDecisionDTO convert(VoteDecision voteDecision) {
+            VoteDecisionDTO dtoDecision = new VoteDecisionDTO();
+            dtoDecision.setOptionId(voteDecision.getVoteOption().getId());
+            dtoDecision.setParticipantId(voteDecision.getVoteParticipant().getId());
+            
+            if (voteDecision.getDecision() != null) {
+                dtoDecision.setDecision(voteDecision.getDecision().getValue());
+            } else if (voteDecision.getParticipants() != null) {
+                dtoDecision.setParticipants(voteDecision.getParticipants());
+            }
+            return dtoDecision;
+        }
+    };
 
-	public final Converter<VoteDecisionDTO, VoteDecision> toDAO = new AbstractConverter<VoteDecisionDTO, VoteDecision>() {
-		@Override
-		protected VoteDecision convert(VoteDecisionDTO dtoDecision) {
-			VoteDecision decision = new VoteDecision();
-			VoteOption voteOption = voteOptionRepository.findOne(dtoDecision.getOptionId());
-			decision.setVoteOption(voteOption);
-			
-			if (dtoDecision.getDecision() != null) {
-				decision.setDecision(Decision.getDecisionById(dtoDecision.getDecision()));				
-			} else if (dtoDecision.getParticipants() != null) {
-				decision.setParticipants(dtoDecision.getParticipants());				
-			}
-			return decision;
-		}
-	};
+	public final Converter<VoteDecisionDTO, VoteDecision> toDAO = new AbstractConverter<>() {
+        @Override
+        protected VoteDecision convert(VoteDecisionDTO dtoDecision) {
+            VoteDecision decision = new VoteDecision();
+            VoteOption voteOption = voteOptionRepository.findById(dtoDecision.getOptionId()).orElse(null);
+            decision.setVoteOption(voteOption);
+            
+            if (dtoDecision.getDecision() != null) {
+                decision.setDecision(Decision.getDecisionById(dtoDecision.getDecision()));
+            } else if (dtoDecision.getParticipants() != null) {
+                decision.setParticipants(dtoDecision.getParticipants());
+            }
+            return decision;
+        }
+    };
 }
