@@ -9,7 +9,6 @@ import io.jsonwebtoken.*;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Matchers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
@@ -26,7 +25,7 @@ import java.util.Date;
 
 import static de.iks.rataplan.testutils.TestConstants.AUTHUSER_1;
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 @ActiveProfiles("test")
 @RunWith(SpringJUnit4ClassRunner.class)
@@ -50,8 +49,11 @@ public class AuthServiceTest {
         KeyPairGenerator kpg = KeyPairGenerator.getInstance("RSA");
         kpg.initialize(2048);
         keyPair = kpg.generateKeyPair();
-        when(keyResolver.resolveSigningKey(Matchers.isA(JwsHeader.class), Matchers.isA(Claims.class))).thenReturn(
-            keyPair.getPublic());
+        when(keyResolver.resolveSigningKey(
+            isA(JwsHeader.class),
+            isA(Claims.class)
+        ))
+            .thenReturn(keyPair.getPublic());
     }
     
     @Test
